Refactor: Entity 수정, 폴더링 변경
